requirements.txt
# requirements.txt
streamlit
pandas
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Pillow
pytz
streamlit-geolocation  
geopy                  